backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
psycopg2-binary==2.9.7
pgvector==0.2.4
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
mem0ai==0.1.117
litellm==1.77.1
openai>=1.0.0
dj-database-url==3.0.1
whitenoise==6.10.0
gunicorn==23.0.0
pytest==8.4.2
pytest-django==4.11.1
